requirements.txt
# DFU Detection using Vision Transformer - Requirements
# Optimized for Apple M3 with 16GB RAM

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Vision Transformer
timm>=0.9.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning Utilities
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Progress Bars
tqdm>=4.66.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Handling
pandas>=2.0.0

# Configuration
pyyaml>=6.0

# Jupyter Support (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Web Interface (optional)
flask>=2.3.0
flask-cors>=4.0.0

# PDF Generation
reportlab>=4.0.0
PyPDF2>=3.0.0

# MongoDB
pymongo>=4.6.0
dnspython>=2.4.0

# Utilities
python-dotenv>=1.0.0
